# Optimize NormalizeLabel with strings.Builder