--- a/strconv.go
+++ b/strconv.go
@@ -20,10 +20,8 @@
 package otlptranslator
 
 import (
-	"github.com/grafana/regexp"
+	"strings"
 )
-
-var invalidLabelCharRE = regexp.MustCompile(`[^a-zA-Z0-9_]`)
 
 // sanitizeLabelName replaces any characters not valid according to the
 // classical Prometheus label naming scheme with an underscore.
@@ -31,5 +29,14 @@ var invalidLabelCharRE = regexp.MustCompile(`[^a-zA-Z0-9_]`)
 // not starting with a digit 0-9), and hence should only be used if the label
 // name is prefixed with a known valid string.
 func sanitizeLabelName(name string) string {
-	return invalidLabelCharRE.ReplaceAllString(name, "_")
+	var b strings.Builder
+	b.Grow(len(name))
+	for _, r := range name {
+		if (r >= 'a' && r <= 'z') || (r >= 'A' && r <= 'Z') || (r >= '0' && r <= '9') {
+			b.WriteRune(r)
+		} else {
+			b.WriteRune('_')
+		}
+	}
+	return b.String()
 }