@ArthurSens 2025-03-26
Removing regex and using strings.Builder show optimizations in speed, bytes per operation and in number of allocations :)